Refactor Google Analytics
